Changed from using Google Sheets API to Google Drive API.
Changed from using Google Sheets API to Google Drive API.
I did this because I think it'll be better to create the .csv file
locally then uploading the .csv file to Google Drive API to create a
Google Sheet. This will save overhead with HTTP requests since the
appllication does not require constant updates to the Google Sheets.
Updates can be made locally in a Python data structure and converted to
.csv.